requirements.txt
agate==1.9.1
annotated-types==0.7.0
anyio==4.13.0
attrs==26.1.0
babel==2.18.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.7
click==8.3.2
colorama==0.4.6
cryptography==46.0.6
daff==1.4.2
db-dtypes==1.5.1
dbt-adapters==1.22.10
dbt-bigquery==1.11.1
dbt-common==1.37.3
dbt-core==1.11.7
dbt-extractor==0.6.0
dbt-protos==1.0.443
dbt-semantic-interfaces==0.9.0
deepdiff==8.6.2
distro==1.9.0
docstring_parser==0.17.0
fastjsonschema==2.21.2
google-api-core==2.30.2
google-auth==2.49.1
google-auth-oauthlib==1.3.1
google-cloud-aiplatform==1.145.0
google-cloud-bigquery==3.41.0
google-cloud-core==2.5.1
google-cloud-dataproc==5.26.0
google-cloud-resource-manager==1.17.0
google-cloud-storage==3.1.1
google-crc32c==1.8.0
google-genai==1.70.0
google-resumable-media==2.8.2
googleapis-common-protos==1.74.0
grpc-google-iam-v1==0.14.4
grpcio==1.80.0
grpcio-status==1.80.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
importlib_metadata==8.9.0
isodate==0.7.2
Jinja2==3.1.6
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
jupyter_core==5.9.1
leather==0.4.1
MarkupSafe==3.0.3
mashumaro==3.14
more-itertools==10.8.0
msgpack==1.1.2
nbformat==5.10.4
networkx==3.6.1
numpy==2.4.4
oauthlib==3.3.1
orderly-set==5.5.0
packaging==26.0
pandas==2.3.3
pandas-gbq==0.34.1
parsedatetime==2.6
pathspec==0.12.1
platformdirs==4.9.4
proto-plus==1.27.2
protobuf==6.33.6
psutil==7.2.2
pyarrow==23.0.1
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
pydata-google-auth==1.9.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-slugify==8.0.4
pytimeparse==1.1.8
pytz==2026.1.post1
PyYAML==6.0.3
referencing==0.37.0
requests==2.33.1
requests-oauthlib==2.0.0
rpds-py==0.30.0
setuptools==82.0.1
six==1.17.0
sniffio==1.3.1
snowplow-tracker==1.1.0
sqlparse==0.5.4
tenacity==9.1.4
text-unidecode==1.3
traitlets==5.14.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2026.1
urllib3==2.6.3
websockets==16.0
zipp==3.23.0
